Refactor card components: Update DownloadCard import paths and integrate new cardDownload.vue into sections for enhanced functionality and layout consistency.
