+++ src/config/helperCreateSection.js
@@ -7,12 +7,14 @@ import Panel from '@/components/cards-new/panel-config/SectionCards.vue';
 import CardData from '@/components/cards-new/cardData.vue';
 import CardInfo from '@/components/cards-new/cardInfo.vue';
 import CardRanking from '@/components/cards-new/cardRanking.vue';
+import CardDownload from '@/components/cards-new/cardDownload.vue';
 
 // Mapa de componentes para facilitar o acesso
 const COMMON_COMPONENTS = {
   CardData,
   CardInfo,
-  CardRanking
+  CardRanking,
+  CardDownload
 };
 
 /**
